agent-evaluation-lab/agents/researcher/pyproject.toml
[project]
name = "researcher"
version = "0.1.0"
dependencies = [
    "google-adk",
    "wikipedia",
    "uvicorn",
    "fastapi"
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
